Fix header include order for building QvE
This was a hiding bug since long that came out as part of adding a new structure in the code. The order of the ENCLAVE_INC_PATH had $(COMMON_INCLUDE) at the beginning that shadowed the local sgx_quote_5.h (and probably other similar headers) as they got fixed from SGX SDK installation directory or /usr/include.

Changing this order helped pick the local headers that are being modified.

---------

Signed-off-by: Sankaranarayanan Venkatasubramanian <sankaranarayanan.venkatasubramanian@intel.com>